Rename dot-transform functions
